Extract Envoy client
